🐛 修复中医健康平台关键Bug集合
核心修复：
• ✅ 端口配置：修复前端Vite代理从8001→8000匹配后端
• ✅ 用户状态：创建统一Pinia用户状态管理，替换硬编码用户ID
• ✅ 认证系统：新增统一认证工具authFetch和checkAuth
• ✅ 导入错误：修复所有ElMessage导入缺失问题
• ✅ 数据精度：Float→Numeric(10,2)解决金额计算精度问题
• ✅ 路由守卫：统一权限验证逻辑
• ✅ 生产配置：更新nginx域名配置为tcmlife.top

🎯 解决问题：
- 所有API调用失败（端口不匹配）
- 用户登录注册流程断裂
- 课程购买支付流程异常
- JavaScript运行时错误
- 金额计算精度丢失

📁 主要文件：
- frontend/vite.config.ts
- frontend/src/stores/user.ts
- frontend/src/utils/auth.ts
- frontend/src/views/*.vue
- backend/app/models/*.py

🧪 测试状态：本地环境全部修复完成，等待部署验证

+++ backend/app/models/course.py
@@ -1,7 +1,7 @@
 """
 课程模型
 """
-from sqlalchemy import Column, Integer, String, Text, DateTime, Float, Boolean, Enum, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, DateTime, Float, Boolean, Enum, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database import Base
@@ -29,7 +29,7 @@ class Course(Base):
     description = Column(Text)
     category = Column(Enum(CourseCategory), nullable=False)
     duration = Column(String)  # 例如: "10课时"
-    price = Column(Float)
+    price = Column(Numeric(8, 2))
     image_url = Column(String)
     is_free = Column(Boolean, default=False)
     is_published = Column(Boolean, default=False)
